refactor: restructured code - added src folder
