+++ swagger/index.js
@@ -1,7 +1,7 @@
 const swaggerAutogen = require('swagger-autogen');
 
 const outputFile = './swagger/swagger_output.json';
-const endpointsFiles = ['./app.js'];
+const endpointsFiles = ['./src/app.js'];
 
 const doc = {
     info: {
